test: added tests for full missile save/load

+++ src/main/java/icbm/classic/prefab/entity/EntityICBM.java
@@ -33,17 +33,17 @@ public abstract class EntityICBM extends Entity implements IWorldPosition
     @Override
     protected void entityInit()
     {
-        this.dataManager.register(HEALTH, Float.valueOf(0));
+        this.dataManager.register(HEALTH, (float) 0);
     }
 
     public float getHealth()
     {
-        return ((Float)this.dataManager.get(HEALTH)).floatValue();
+        return this.dataManager.get(HEALTH);
     }
 
     public void setHealth(float health)
     {
-        this.dataManager.set(HEALTH, Float.valueOf(MathHelper.clamp(health, 0.0F, this.getMaxHealth())));
+        this.dataManager.set(HEALTH, MathHelper.clamp(health, 0.0F, this.getMaxHealth()));
     }
 
     public float getMaxHealth()
